requirements.txt
click==7.1.2
h5py==3.1.0
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
progressbar33==2.4
pylsl==1.13.6
PyQt5==5.15.1
pyqtgraph==0.11.0
scikit_learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
tensorflow==2.3.1
